+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Updated `EuiText`s `blockquote` font-size/line-height to match the base font-size/line-height which is the same as paragraphs ([#4663](https://github.com/elastic/eui/pull/4663))
 - Added `markdownFormatProps` prop to `EuiMarkdownEditor` to extend the props passed to the rendered `EuiMarkdownFormat` ([#4663](https://github.com/elastic/eui/pull/4663))
 - Updated `barSeriesStyle.displayValue` of the elastic charts `Theme` to allow `textInvertible` and `textContrast` enabled by default ([#4845](https://github.com/elastic/eui/pull/4845))
+- Added optional virtualized line rendering to `EuiCodeBlock` ([#4952](https://github.com/elastic/eui/pull/4952))
 
 ## [`36.0.0`](https://github.com/elastic/eui/tree/v36.0.0)